separate twins functions
